Move Langevin Dynamics to dedicated header file
